Refactor modnotes: separate cog (view), model, controller
