+++ config.example.json
@@ -43,8 +43,5 @@
       ["url1", "artist name 1"],
       ["url2", "artist name 2"]
     ]
-  },
-  "modnotes": {
-
   }
 }
